requirement.txt
selenium
pytest
allure-pytest
html-testRunner
openpyxl
pytest-ordering
pytest-xdist
PyAutoIt
webdriver-manager